requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==2.4.0
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
